requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-swagger==2.2.0
django-sendgrid-v5==0.9.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sendgrid==6.4.7
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
uritemplate==3.0.1
urllib3==1.25.10
